change Description Sidebar Opening to View Button.
Description :
Earlier you click anywhere in the table row and sidebar which contains the description of meeting gets opened. Issue with this was that if you click on delete button then indirectly you clicked the table row too so after the deleteion of that row, description sidebar get's opened with the details of deleted row meeting, which was a bug and not a feature.

Remaining Work :
 - Add functionality to edit any field in an exixting meeting within the sidebar or use a modal.